moved network mismatchbanner to comps
