+++ packages/simplified/src/modules/common/labels.styles.less
@@ -187,61 +187,3 @@
     }
   }
 }
-
-.NetworkMismatch {
-  .ThemeBorder;
-  .text-16-bold;
-
-  border: none;
-  background: rgba(var(--failed-rgb), 0.15);
-  color: var(--failed);
-  text-align: center;
-  width: 100%;
-  flex: 1;
-  display: flex;
-  flex-flow: row nowrap;
-  align-items: center;
-  justify-content: center;
-  padding: @size-13 @size-16;
-  max-height: @size-48;
-  min-height: @size-48;
-  height: @size-48;
-  margin-bottom: @size-16;
-  margin-top: -@size-24;
-  white-space: nowrap;
-
-  @media @breakpoint-mobile-tablet {
-    .text-14-semi-bold;
-
-    margin-top: 0;
-  }
-
-  &.Market {
-    margin-bottom: -@size-24;
-
-    @media @breakpoint-mobile-tablet {
-      margin-bottom: @size-16;
-    }
-  }
-}
-
-.WarningBanner {
-  background: rgba(var(--warning-rgb), 0.15);
-  color: var(--warning);
-  word-break: normal;
-  word-wrap: normal;
-  white-space: normal;
-  display: flex;
-  flex-flow: row nowrap;
-
-  > a {
-    margin-left: @size-4;
-    text-decoration: underline;
-  }
-
-  @media @breakpoint-mobile {
-    display: inline;
-    height: @size-60;
-    max-height: @size-60;
-  }
-}
